wip: fix delete grace logic

+++ cmd/main.go
@@ -12,7 +12,7 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-const version string = "2.0.0-rc3"
+const version string = "2.0.0-rc4"
 
 func main() {
 	log.SetOutput(os.Stdout)
@@ -63,7 +63,10 @@ func main() {
 			if config.Cloudflare.DeleteGrace == 0 {
 				internal.DeleteCloudflareDNSRecord(orphanedRecord)
 				for _, deleteGraceRecord := range deleteGraceRecords {
-					log.Infof("Cleaning up delete grace DNS record for %s", strings.TrimPrefix(deleteGraceRecord.Name, "_syncflaer._deletegrace."))
+					if !strings.Contains(deleteGraceRecord.Name, orphanedRecord.Name) {
+						continue
+					}
+					log.Infof("Cleaning up delete grace DNS record for %s", orphanedRecord.Name)
 					internal.DeleteCloudflareDNSRecord(deleteGraceRecord)
 				}
 				continue
@@ -78,7 +81,7 @@ func main() {
 				deleteGrace -= 1
 				if deleteGrace > 0 {
 					deleteGraceRecord.Content = strconv.Itoa(deleteGrace)
-					log.Infof("Waiting %d more runs until deleting DNS record %s", deleteGrace-1, orphanedRecord.Name)
+					log.Infof("Waiting %d more runs until deleting DNS record %s", deleteGrace, orphanedRecord.Name)
 					internal.UpdateCloudflareDNSRecord(deleteGraceRecord)
 					continue
 				}
@@ -91,13 +94,21 @@ func main() {
 					Name:    fmt.Sprintf("_syncflaer._deletegrace.%s", orphanedRecord.Name),
 					Content: strconv.Itoa(config.Cloudflare.DeleteGrace),
 				}
-				log.Infof("Waiting %d more runs until deleting DNS record %s", config.Cloudflare.DeleteGrace-1, orphanedRecord.Name)
+				log.Infof("Waiting %d more runs until deleting DNS record %s", config.Cloudflare.DeleteGrace, orphanedRecord.Name)
 				internal.CreateCloudflareDNSRecord(deleteGraceRecord)
 			}
 		}
 	} else {
+	deleteGraceRecord:
 		for _, deleteGraceRecord := range deleteGraceRecords {
-			log.Infof("DNS record %s is not orphaned anymore", strings.TrimPrefix(deleteGraceRecord.Name, "_syncflaer._deletegrace."))
+			for _, cloudflareDNSRecord := range cloudflareDNSRecords {
+				if strings.Contains(deleteGraceRecord.Name, cloudflareDNSRecord.Name) && cloudflareDNSRecord.Name != config.Cloudflare.ZoneName {
+					log.Infof("DNS record %s is not orphaned anymore", cloudflareDNSRecord.Name)
+					internal.DeleteCloudflareDNSRecord(deleteGraceRecord)
+					continue deleteGraceRecord
+				}
+			}
+			log.Infof("Cleaning up unused delete grace DNS record %s", deleteGraceRecord.Name)
 			internal.DeleteCloudflareDNSRecord(deleteGraceRecord)
 		}
 		log.Debug("No orphaned DNS records")
